requirements.txt
# Core dependencies - versões compatíveis
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.8.0,<3.0.0
python-dotenv>=1.0.0

# AI/ML dependencies - versões flexíveis
anthropic>=0.34.0
chromadb>=0.4.0
sentence-transformers>=2.0.0
transformers>=4.20.0

# Community MCP dependencies
mitreattack-python>=4.0.0
mcp>=1.11.0

# Data processing
pandas>=2.0.0
numpy>=1.20.0
requests>=2.25.0
beautifulsoup4>=4.10.0
pypdf2>=3.0.0

# Database and logging
sqlalchemy>=2.0.0
alembic>=1.10.0
psycopg2-binary>=2.9.0
asyncpg>=0.29.0
aiosqlite>=0.19.0
motor>=3.3.0
loguru>=0.7.0

# Frontend
streamlit>=1.25.0
plotly>=5.15.0

